feat: Add brush size change and sand destruction functionality
The code changes introduce the ability to change the brush size using the mouse wheel and destroy sand under the mouse position using the space bar. This enhances the user's control over the sand simulation.

feat: Add stickness to sand simulation. Now the sand are not updated if they are in REST state, but if they are adjacent to a empty cell they become FREEFALL or SLIDING.